move to legacy: usage_analytics, weave_inspector
