+++ weave/legacy/weave_inspector.py
@@ -107,8 +107,8 @@ import tabulate
 
 from weave.legacy import graph
 
-from . import weave_types as types
-from .legacy.partial_object import PartialObjectType
+from .. import weave_types as types
+from .partial_object import PartialObjectType
 
 
 def _trimmed_string(s: str, max_len: int = 20) -> str:
