+++ weave/query_api.py
@@ -29,7 +29,7 @@ from . import types_numpy as _types_numpy
 from . import errors
 from weave.legacy.decorators import weave_class, mutation, type
 
-from . import usage_analytics
+from weave.legacy import usage_analytics
 from weave.legacy.context import (
     use_fixed_server_port,
     use_frontend_devmode,
